+++ koopman_autoencoder/models/autoencoder.py
@@ -84,46 +84,3 @@ class KoopmanAutoencoder(nn.Module):
             preds.append(self.decode(z))
 
         return preds
-
-
-class KoopmanLoss(nn.Module):
-    def __init__(self, alpha=1.0, beta=0.1, rollout_steps=1):
-        super(KoopmanLoss, self).__init__()
-        self.alpha = alpha  # weight for prediction loss
-        self.beta = beta    # weight for latent loss
-        self.rollout_steps = rollout_steps
-
-    def forward(self, x_recon, x_preds, latent_pred_differences, x_true, x_future):
-        recon_loss = F.mse_loss(x_recon, x_true)
-
-        pred_loss = 0.0
-        latent_loss = 0.0
-        for t in range(self.rollout_steps):
-            pred_loss += F.mse_loss(x_preds[t], x_future[:, t])
-            latent_loss += F.mse_loss(latent_pred_differences[t], torch.zeros_like(latent_pred_differences[t]))
-
-        pred_loss /= self.rollout_steps
-        latent_loss /= self.rollout_steps
-
-        total_loss = recon_loss + self.alpha * pred_loss + self.beta * latent_loss
-        return total_loss, recon_loss, pred_loss, latent_loss
-
-def train_koopman_autoencoder(model, train_loader, optimizer, criterion, device, num_epochs=100, rollout_steps=3):
-    model.train()
-    for epoch in range(num_epochs):
-        total_loss = 0
-        for batch_idx, (x, x_future) in enumerate(train_loader):
-            x = x.to(device)
-            x_future = x_future.to(device)  # shape: [B, rollout_steps, C, H, W]
-
-            optimizer.zero_grad()
-            x_recon, x_preds, z_preds = model(x, rollout_steps=rollout_steps)
-            loss, recon_loss, pred_loss, latent_loss = criterion(x_recon, x_preds, z_preds, x, x_future)
-
-            loss.backward()
-            optimizer.step()
-            total_loss += loss.item()
-
-        avg_loss = total_loss / len(train_loader)
-        if (epoch + 1) % 10 == 0:
-            print(f'Epoch [{epoch+1}/{num_epochs}], Loss: {avg_loss:.4f}, Recon: {recon_loss:.4f}, Pred: {pred_loss:.4f}, Latent: {latent_loss:.4f}')
