+++ koopman_autoencoder/models/eval.py
@@ -1,0 +1,26 @@
+import torch
+
+def evaluate_model(model, dataloader, criterion, device):
+    model.eval()
+    total_loss = 0
+    total_recon_loss = 0
+    total_pred_loss = 0
+    total_latent_loss = 0
+
+    with torch.no_grad():
+        for x, x_next_seq in dataloader:
+            x, x_next_seq = x.to(device), x_next_seq.to(device)
+            x_recon, x_preds, z_preds, latent_pred_differences = model(x, rollout_steps=x_next_seq.size(1))
+            loss, recon_loss, pred_loss, latent_loss = criterion(x_recon, x_preds, latent_pred_differences, x, x_next_seq)
+            total_loss += loss.item()
+            total_recon_loss += recon_loss.item()
+            total_pred_loss += pred_loss.item()
+            total_latent_loss += latent_loss.item()
+
+    n_batches = len(dataloader)
+    return {
+        'loss': total_loss / n_batches,
+        'recon_loss': total_recon_loss / n_batches,
+        'pred_loss': total_pred_loss / n_batches,
+        'latent_loss': total_latent_loss / n_batches
+    }
